# mydentist/static/swagger/settings.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API Documentation for MyDentist Platform",
        "title": "MyDentist API"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/results/": {
            "post": {
                "tags": [
                    "base"
                ],
                "summary": "Results page of MyDentist Platform",
                "description": "",
                "operationId": "results",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Results"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/ResultsResults"
                        }
                    },
                    "400": {
                        "description": "No data"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/dentist/": {
            "post": {
                "tags": [
                    "base"
                ],
                "summary": "Dentist page of MyDentist Platform",
                "description": "",
                "operationId": "dentist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Parameters",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dentist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/DentistResult"
                        }
                    },
                    "400": {
                        "description": "No data"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
...
        "/notifications/": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Notifications API of MyDentist Platform",
                "description": "",
                "operationId": "notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "400": {
                        "description": "No data"
                    },
                    "401": {
                        "description": "Token Error"
                    }
                }
            }
        },
        "/settings/": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Settings page of MyDentist Platform",
                "description": "",
                "operationId": "settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/SettingsResult"
                        }
                    },
                    "400": {
                        "description": "No data"
                    },
                    "401": {
                        "description": "Token Error"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "Results": {
            "type": "object",
            "required": [
                "language",
                "region",
                "service",
                "sort_by"
            ],
            "properties": {
                "language": {
                    "type": "string",
                    "example": "uz"
                },
                "region": {
                    "type": "integer",
                    "format": "int64",
                    "example": 5
                },
                "service": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "woman": {
                    "type": "boolean",
                    "default": false
                },
                "hour_24": {
                    "type": "boolean",
                    "default": false
                },
                "no_que": {
                    "type": "boolean",
                    "default": false
                },
                "lat": {
                    "type": "number",
                    "default": 40.808288
                },
                "long": {
                    "type": "number",
                    "default": 72.344401
                },
                "sort_by": {
                    "type": "string",
                    "example": "price",
                    "enum": [
                        "price",
                        "near"
                    ]
                }
            }
        },
        "ResultsResults": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/ResultsResult"
            }
        },
        "Dentist": {
            "type": "object",
            "required": [
                "language",
                "slug"
            ],
            "properties": {
                "language": {
                    "type": "string",
                    "example": "uz"
                },
                "slug": {
                    "type": "string",
                    "example": "qodirov-muhammadali-mahammadsoli-ogli"
                }
            }
        },
        "DentistResult": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string",
                    "example": "Qodirov Muhammadali Mahammadsoli o'g'li"
                },
                "image": {
                    "type": "string",
                    "example": "/media/dentists/photos/qodirov-muhammadali-mahammadsoli-ogli.jpg"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+998901709993"
                },
                "clinic_name": {
                    "type": "string",
                    "example": "Sog'lom tabassum"
                },
                "address": {
                    "type": "string",
                    "example": "Bobur shoh ko'chasi, 109-b uy"
                },
                "orientir": {
                    "type": "string",
                    "example": "Hamkor biznes markazi"
                },
                "latitude": {
                    "type": "number",
                    "example": 40.784638
                },
                "longitude": {
                    "type": "number",
                    "example": 72.332524
                },
                "worktime_begin": {
                    "type": "string",
                    "example": "8:00"
                },
                "worktime_end": {
                    "type": "string",
                    "example": "17:15"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DentistService"
                    }
                },
                "cabinet_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CabinetImage"
                    }
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Reason"
                    }
                }
            }
        },
...
        "Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "SettingsResult": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "object",
                    "properties": {
                        "first_name": {
                            "type": "string",
                            "example": "Abror"
                        },
                        "last_name": {
                            "type": "string",
                            "example": "Mamataliev"
                        },
                        "gender": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1,
                            "enum": [1,2]
                        },
                        "birthday": {
                            "type": "string",
                            "example": "26.04.2002",
                            "enum": [
                                "dd.mm.YYYY"
                            ]
                        },
                        "email": {
                            "type": "string",
                            "example": ""
                        },
                        "phone_number": {
                            "type": "string",
                            "example": "(90) 346-26-04"
                        },
                        "address": {
                            "type": "string",
                            "example": "Toshkent"
                        },
                        "language": {
                            "type": "string",
                            "example": "uz"
                        }
                    }
                },
                "illness": {
                    "type": "object",
                    "properties": {
                        "diabet": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "integer",
                                    "format": "int64",
                                    "example": 1
                                },
                                "desc": {
                                    "type": "string",
                                    "example": "Yo'q"
                                }
                            }
                        }
                    }
                },
                "otherillness": {
                    "type": "object",
                    "properties": {
                        "epilepsy": {
                            "type": "object",
                            "example": null
                        }
                    }
                }
            }
        },
...
        "ResultsResult": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string",
                    "example": "/media/dentists/photos/qodirov-muhammadali-mahammadsoli-ogli.jpg"
                },
                "clinic_name": {
                    "type": "string",
                    "example": "Sog'lom tabassum"
                },
                "fullname": {
                    "type": "string",
                    "example": "Qodirov Muhammadali Mahammadsoli o'g'li"
                },
                "address": {
                    "type": "string",
                    "example": "Bobur shoh ko'chasi, 109-b uy"
                },
                "orientir": {
                    "type": "string",
                    "example": "Hamkor biznes markazi"
                },
                "latitude": {
                    "type": "number",
                    "example": 40.784638
                },
                "longitude": {
                    "type": "number",
                    "example": 72.332524
                },
                "worktime_begin": {
                    "type": "string",
                    "example": "8:00"
                },
                "worktime_end": {
                    "type": "string",
                    "example": "17:15"
                },
                "is_fullday": {
                    "type": "boolean",
                    "example": false
                },
                "phone_number": {
                    "type": "string",
                    "example": "+998901709993"
                },
                "service_name": {
                    "type": "string",
                    "example": "Konsultatsiya"
                },
                "service_price_min": {
                    "type": "integer",
                    "format": "int64",
                    "example": 0
                },
                "service_price_max": {
                    "type": "integer",
                    "format": "int64",
                    "example": 50000
                },
                "slug": {
                    "type": "string",
                    "example": "qodirov-muhammadali-mahammadsoli-ogli"
                }
            }
        },
        "DentistService": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Konsultatsiya"
                },
                "price": {
                    "type": "integer",
                    "format": "int64",
                    "example": 50000
                },
                "duration": {
                    "type": "integer",
                    "format": "int64",
                    "example": 15
                }
            }
        },
        "CabinetImage": {
            "type": "string",
            "example": "/media/dentists/cabinet_photos/qodirov-muhammadali-mahammadsoli-ogli-1.jpg"
        },
...
    }
}